Address/DB Service: Optimize indexes for queries
Being able to query for previous output information to determine the address
and the satoshis amount will be useful for updating address sums indexes when
a new block is connected to the chain. Furthermore this may help with calculating
the fees for a transaction and other such information needed for a transaction.

Using and index to determine if an output is spent instead of calling bitcoin
should have performance benefits, as this query has shown to be expensive for
concurrency.

Updating the block index to include the block hash, transaction count and the
total bytes for the block will avoid needing to query 20 plus blocks to determine
a summary of the latest blocks in the chain.

This commit only changes the indexing and queries still need to be updated
to reflect the changes.